[#16] Fix escape key binding
We were previously registering the escape key globally, leaving it unavailable for other apps to leverage. This commit moves the key binding into the panes themselves, rather than binding it in the main process.
