+++ panes/about.html
@@ -25,6 +25,7 @@
 
     <small>Copyright &copy; 2017 Trezy.com</small>
 
+    <script src="pane.js"></script>
     <script src="about.js"></script>
   </body>
 </html>
