+++ panes/preferences.html
@@ -45,6 +45,7 @@
       </table>
     </form>
 
+    <script src="pane.js"></script>
     <script src="preferences.js"></script>
   </body>
 </html>
